chore: combine some components (#60)
