+++ bases/renku_data_services/data_api/app.py
@@ -2,8 +2,9 @@
 from sanic import Sanic
 
 from renku_data_services.base_api.error_handler import CustomErrorHandler
-from renku_data_services.base_blueprints import MiscBP
-from renku_data_services.crc_blueprints import (
+from renku_data_services.base_api.misc import MiscBP
+from renku_data_services.crc import apispec
+from renku_data_services.crc.blueprints import (
     ClassesBP,
     QuotaBP,
     ResourcePoolsBP,
@@ -11,9 +12,8 @@ from renku_data_services.crc_blueprints import (
     UserResourcePoolsBP,
     UsersBP,
 )
-from renku_data_services.crc_schemas import apispec
 from renku_data_services.data_api.config import Config
-from renku_data_services.storage_blueprints import StorageBP, StorageSchemaBP
+from renku_data_services.storage.blueprints import StorageBP, StorageSchemaBP
 
 
 def register_all_handlers(app: Sanic, config: Config) -> Sanic:
