+++ bases/renku_data_services/data_api/config.py
@@ -11,10 +11,11 @@ from tenacity import retry, stop_after_attempt, stop_after_delay, wait_fixed
 from yaml import safe_load
 
 import renku_data_services.base_models as base_models
-import renku_data_services.crc_schemas
-import renku_data_services.resource_pool_models as models
-import renku_data_services.storage_schemas
+import renku_data_services.crc
+import renku_data_services.storage
 from renku_data_services import errors
+from renku_data_services.crc import models
+from renku_data_services.crc.db import ResourcePoolRepository, UserRepository
 from renku_data_services.data_api.server_options import (
     ServerOptions,
     ServerOptionsDefaults,
@@ -24,8 +25,7 @@ from renku_data_services.git.gitlab import DummyGitlabAPI, GitlabAPI
 from renku_data_services.k8s.clients import DummyCoreClient, DummySchedulingClient, K8sCoreClient, K8sSchedulingClient
 from renku_data_services.k8s.quota import QuotaRepository
 from renku_data_services.migrations.core import DataRepository
-from renku_data_services.resource_pool_adapters import ResourcePoolRepository, UserRepository
-from renku_data_services.storage_adapters import StorageRepository
+from renku_data_services.storage.db import StorageRepository
 from renku_data_services.users.dummy import DummyAuthenticator, DummyUserStore
 from renku_data_services.users.gitlab import GitlabAuthenticator
 from renku_data_services.users.keycloak import KcUserStore, KeycloakAuthenticator
@@ -87,11 +87,11 @@ class Config:
     server_defaults_file: Optional[str] = None
 
     def __post_init__(self):
-        spec_file = Path(renku_data_services.crc_schemas.__file__).resolve().parent / "api.spec.yaml"
+        spec_file = Path(renku_data_services.crc.__file__).resolve().parent / "api.spec.yaml"
         with open(spec_file, "r") as f:
             crc_spec = safe_load(f)
 
-        spec_file = Path(renku_data_services.storage_schemas.__file__).resolve().parent / "api.spec.yaml"
+        spec_file = Path(renku_data_services.storage.__file__).resolve().parent / "api.spec.yaml"
         with open(spec_file, "r") as f:
             storage_spec = safe_load(f)
