--- components/renku_data_services/base_blueprints/__init__.py
@@ -1,2 +1,0 @@
-"""Common Blueprints."""
-from renku_data_services.base_blueprints.core import *  # noqa: F401, F403
