+++ components/renku_data_services/crc/__init__.py
@@ -1,0 +1,1 @@
+"""Blueprints for CRC."""
