+++ components/renku_data_services/crc/apispec.py
@@ -1,6 +1,6 @@
 # generated by datamodel-codegen:
 #   filename:  api.spec.yaml
-#   timestamp: 2023-09-11T20:25:56+00:00
+#   timestamp: 2023-10-24T00:52:08+00:00
 
 from __future__ import annotations
 
@@ -8,7 +8,7 @@ from typing import List, Optional
 
 from pydantic import ConfigDict, Field, RootModel
 
-from renku_data_services.crc_schemas.base import BaseAPISpec
+from renku_data_services.crc.apispec_base import BaseAPISpec
 
 
 class Version(BaseAPISpec):
