+++ components/renku_data_services/crc/orm.py
@@ -6,7 +6,7 @@ from sqlalchemy.orm import DeclarativeBase, Mapped, MappedAsDataclass, mapped_co
 from sqlalchemy.schema import ForeignKey
 
 import renku_data_services.base_models as base_models
-import renku_data_services.resource_pool_models as models
+from renku_data_services.crc import models
 
 metadata_obj = MetaData(schema="resource_pools")  # Has to match alembic ini section name
