--- components/renku_data_services/crc_blueprints/__init__.py
@@ -1,2 +1,0 @@
-"""Blueprints for CRC."""
-from renku_data_services.crc_blueprints.core import *  # noqa: F401, F403
