+++ components/renku_data_services/crc_migrations/versions/024b215f9a14_default_pool_access_control.py
@@ -5,10 +5,9 @@ Revises: 3097ca05ce65
 Create Date: 2023-1--15 12:33:45.53902
 
 """
-from alembic import op
 import sqlalchemy as sa
+from alembic import op
 from sqlalchemy.sql import expression
-
 
 # revision identifiers, used by Alembic.
 revision = "024b215f9a14"
