--- components/renku_data_services/crc_schemas/__init__.py
@@ -1,1 +1,0 @@
-"""Various schemas used by the crc service."""
