--- components/renku_data_services/resource_pool_adapters/__init__.py
@@ -1,2 +1,0 @@
-"""Database adapters for resource pools."""
-from renku_data_services.resource_pool_adapters.adapter import *  # noqa: F401, F403
