--- components/renku_data_services/resource_pool_models/__init__.py
@@ -1,2 +1,0 @@
-"""Domain models and exceptions."""
-from renku_data_services.resource_pool_models.core import *  # noqa: F401, F403
