+++ components/renku_data_services/storage/__init__.py
@@ -1,0 +1,1 @@
+"""Blueprints for cloud storage."""
