+++ components/renku_data_services/storage/api.spec.yaml
@@ -167,7 +167,6 @@ components:
         - type: string
         - type: boolean
         - type: object
-        - type: 'null'
     CloudStorageUrl:
       allOf:
         - $ref: "#/components/schemas/GitRequest"
