+++ components/renku_data_services/storage/apispec.py
@@ -1,6 +1,6 @@
 # generated by datamodel-codegen:
 #   filename:  api.spec.yaml
-#   timestamp: 2023-09-25T07:29:48+00:00
+#   timestamp: 2023-10-24T00:52:09+00:00
 
 from __future__ import annotations
 
@@ -9,7 +9,7 @@ from typing import Any, Dict, List, Optional, Union
 
 from pydantic import Field, RootModel
 
-from renku_data_services.storage_schemas.base import BaseAPISpec
+from renku_data_services.storage.apispec_base import BaseAPISpec
 
 
 class Example(BaseAPISpec):
