+++ components/renku_data_services/storage/db.py
@@ -5,9 +5,9 @@ from sqlalchemy.ext.asyncio import AsyncSession, create_async_engine
 from sqlalchemy.orm import sessionmaker
 
 import renku_data_services.base_models as base_models
-import renku_data_services.storage_models as models
 from renku_data_services import errors
-from renku_data_services.storage_adapters import schemas
+from renku_data_services.storage import models
+from renku_data_services.storage import orm as schemas
 
 
 class _Base:
