--- components/renku_data_services/storage_adapters/__init__.py
@@ -1,2 +1,0 @@
-"""Database adapters for cloud storage data."""
-from renku_data_services.storage_adapters.core import *  # noqa: F401, F403
