--- components/renku_data_services/storage_blueprints/__init__.py
@@ -1,2 +1,0 @@
-"""Blueprints for cloud storage."""
-from renku_data_services.storage_blueprints.core import *  # noqa: F401, F403
