+++ components/renku_data_services/storage_migrations/env.py
@@ -1,6 +1,6 @@
 """Alembic setup and environment used for database migrations."""
 from renku_data_services.migrations.env import run_migrations
-from renku_data_services.storage_adapters.schemas import BaseORM
+from renku_data_services.storage.orm import BaseORM
 
 target_metadata = BaseORM.metadata
 run_migrations(target_metadata)
