--- components/renku_data_services/storage_models/__init__.py
@@ -1,2 +1,0 @@
-"""Database models for cloud storage."""
-from renku_data_services.storage_models.core import *  # noqa: F401, F403
