--- components/renku_data_services/storage_schemas/__init__.py
@@ -1,4 +1,0 @@
-"""Apispec schemas for storage service."""
-from renku_data_services.storage_schemas import core
-
-__all__ = ["core"]
