+++ projects/renku_data_service/pyproject.toml
@@ -9,21 +9,14 @@ packages = [
   { include = "renku_data_services/data_api", from = "../../bases" },
   { include = "renku_data_services/migrations", from = "../../components" },
   { include = "renku_data_services/storage_migrations", from = "../../components" },
-  { include = "renku_data_services/storage_schemas", from = "../../components" },
-  { include = "renku_data_services/storage_adapters", from = "../../components" },
+  { include = "renku_data_services/storage", from = "../../components" },
   { include = "renku_data_services/users", from = "../../components" },
   { include = "renku_data_services/errors", from = "../../components" },
   { include = "renku_data_services/base_models", from = "../../components" },
   { include = "renku_data_services/base_api", from = "../../components" },
-  { include = "renku_data_services/storage_models", from = "../../components" },
   { include = "renku_data_services/crc_migrations", from = "../../components" },
-  { include = "renku_data_services/crc_schemas", from = "../../components" },
+  { include = "renku_data_services/crc", from = "../../components" },
   { include = "renku_data_services/k8s", from = "../../components" },
-  { include = "renku_data_services/resource_pool_adapters", from = "../../components" },
-  { include = "renku_data_services/resource_pool_models", from = "../../components" },
-  { include = "renku_data_services/storage_blueprints", from = "../../components" },
-  { include = "renku_data_services/base_blueprints", from = "../../components" },
-  { include = "renku_data_services/crc_blueprints", from = "../../components" },
   { include = "renku_data_services/utils", from = "../../components" },
   { include = "renku_data_services/git", from = "../../components" },
 ]
