+++ test/bases/renku_data_services/data_api/test_storage.py
@@ -5,13 +5,13 @@ import pytest
 from sanic import Sanic
 from sanic_testing.testing import SanicASGITestClient
 
+from renku_data_services.crc.db import ResourcePoolRepository, UserRepository
 from renku_data_services.data_api.app import register_all_handlers
 from renku_data_services.data_api.config import Config
 from renku_data_services.k8s.clients import DummyCoreClient, DummySchedulingClient
 from renku_data_services.k8s.quota import QuotaRepository
-from renku_data_services.resource_pool_adapters import ResourcePoolRepository, UserRepository
-from renku_data_services.storage_adapters import StorageRepository
-from renku_data_services.storage_schemas.core import RCloneValidator
+from renku_data_services.storage.db import StorageRepository
+from renku_data_services.storage.rclone import RCloneValidator
 from renku_data_services.users.dummy import DummyAuthenticator, DummyUserStore
 
 _valid_storage: dict[str, Any] = {
