+++ test/components/renku_data_services/db/test_sqlalchemy_pool_repo.py
@@ -16,9 +16,9 @@ from hypothesis import strategies as st
 from pydantic import ValidationError
 
 import renku_data_services.base_models as base_models
-import renku_data_services.resource_pool_models as models
 from renku_data_services import errors
-from renku_data_services.resource_pool_adapters import ResourcePoolRepository, UserRepository
+from renku_data_services.crc import models
+from renku_data_services.crc.db import ResourcePoolRepository, UserRepository
 
 
 @given(rp=rp_strat())
