+++ test/components/renku_data_services/k8s/test_k8s_adapter.py
@@ -4,7 +4,7 @@ from test.components.renku_data_services.crc_models.hypothesis import quota_stra
 from hypothesis import given
 from kubernetes import client
 
-import renku_data_services.resource_pool_models as models
+from renku_data_services.crc import models
 from renku_data_services.k8s.clients import DummyCoreClient, DummySchedulingClient
 from renku_data_services.k8s.quota import QuotaRepository
